pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
arrow = "^1.2.1"
python-dateutil = "^2.8.2"
requests = "^2.26.0"
gunicorn = "^20.1.0"
markupsafe = "2.0.1"
pymongo = "^4.2.0"
mongomock = "^4.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pytest-freezegun = "^0.4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
